+++ CHANGELOG.md
@@ -16,6 +16,10 @@
 * Fix bug where files included from included files were being done so
   relative to the working directory of the top-level script. Includes
   should now always be relative to the containing file.
+* Change include behavior so duplicate references and eval bindings
+  are now allowed, and silently overwrite previous bindings. This is
+  needed to support repeated includes (caused by diamond-like include
+  graphs) without using namespacing.
 
 ### 0.1.8
 * Support set forks with `{$id: (foo), (bar)}` syntax. Silent set